% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_cota_maxima_transitividad.R
\name{calcular_cota_maxima_transitividad}
\alias{calcular_cota_maxima_transitividad}
\title{q}
\usage{
calcular_cota_maxima_transitividad(R)
}
\arguments{
\item{R}{}
}
\value{
q
}
\description{
.. content for
}
\details{
.. content for
}
\author{
emilio
}
