% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calcular_cota_maxima_transitividad_matriz_puntos.R
\name{calcular_cota_maxima_transitividad_matriz_puntos}
\alias{calcular_cota_maxima_transitividad_matriz_puntos}
\title{Calculate the upper bound of R with filled with each row of mpoints}
\usage{
calcular_cota_maxima_transitividad_matriz_puntos(R, mpoints)
}
\arguments{
\item{R}{}

\item{mpoints}{a matrix with ncol equals to the missing data. It may contains missing values.}
}
\value{

}
\description{
.. content for
}
\details{
.. content for
}
\author{
emilio
}
