% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_transitividad.R
\name{calcular_transitividad}
\alias{calcular_transitividad}
\title{Calculate the transitivity of a non empty relationship matrix}
\usage{
calcular_transitividad(R)
}
\arguments{
\item{R}{}
}
\value{
the transitivity
}
\description{
.. content for
}
\details{
.. content for
}
\author{
emilio
}
