% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_transitividad_matriz_puntos.R
\name{calcular_transitividad_matriz_puntos}
\alias{calcular_transitividad_matriz_puntos}
\title{Compute the transitivity of R filled with the rows of mpoints}
\usage{
calcular_transitividad_matriz_puntos(R, mpoints)
}
\arguments{
\item{R}{}

\item{mpoints}{a matrix. It cannot have missing data. Number of columns must be the number of missing values in R.}
}
\value{

}
\description{
.. content for
}
\details{
.. content for
}
\author{
emilio
}
