% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chequear_matriz.R
\name{chequear_matriz}
\alias{chequear_matriz}
\title{Check if it is a relation matrix}
\usage{
chequear_matriz(R, verbose = 0L)
}
\arguments{
\item{R}{Matrix}

\item{verbose}{verbose}
}
\value{
a matrix
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for ..
}
\author{
emilio
}
