% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_matriz_aleatoria.R
\name{crear_matriz_aleatoria}
\alias{crear_matriz_aleatoria}
\title{Create a random integer relation matrix}
\usage{
crear_matriz_aleatoria(ndimension, valordiagonal = 10)
}
\arguments{
\item{ndimension}{dimension}

\item{valordiagonal}{the value of the diagonal}
}
\value{
an integer matrix
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for ..
}
\author{
emilio
}
