% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{.onUnload}
\alias{.onUnload}
\title{SIEMPRE PON ESTE FICHERO}
\usage{
.onUnload(libpath)
}
\arguments{
\item{libpath}{1}
}
\value{
1
}
\description{
.. content for description{} (no empty lines) ..
}
\details{
.. content for details{} .. @rawNamespace useDynLib(data.table); useDynLib(emiliocpp)
}
\author{
emilio
}
