% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_texto_matriz.R
\name{leer_texto_matriz}
\alias{leer_texto_matriz}
\title{dd}
\usage{
leer_texto_matriz(texto)
}
\arguments{
\item{texto}{}
}
\value{
dd
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
texto <- "
        [,1] [,2] [,3] [,4] [,5]
   [1,]   10   NA   NA   NA   NA
   [2,]    9   10   NA   NA    6
   [3,]   NA   NA   10   NA   NA
   [4,]   NA   NA   NA   10   NA
   [5,]   NA   NA    4   NA   10
   "
   R <- leer_texto_matriz(texto)
  print(R)
}
\author{
emilio
}
