% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_puntos_canonicos.R
\name{obtener_puntos_canonicos}
\alias{obtener_puntos_canonicos}
\title{Saca los puntos canónicos, representantes de cada clase.}
\usage{
obtener_puntos_canonicos(R, mpoints)
}
\arguments{
\item{R}{una matriz}

\item{mpoints}{un matriz, normalmente la que sale de 'obtener_puntos_optimos(R)'.}
}
\value{
Una matriz con los puntos canónicos.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
1
}
\author{
emilio
}
