% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_puntos_optimos.R
\name{obtener_puntos_optimos}
\alias{obtener_puntos_optimos}
\title{Get a matrix with the solutions}
\usage{
obtener_puntos_optimos(
  R,
  todaslassoluciones = TRUE,
  solovaloresexistentes = TRUE,
  solominimos = TRUE,
  combinacionesmaxi = 10^7,
  verbose = 0L
)
}
\arguments{
\item{R}{matrix}

\item{todaslassoluciones}{TRUE saca todas las soluciones. FALSE: solo saca las primeras que encuentra.}

\item{solovaloresexistentes}{Usa solo los valores que existen en la matriz para rellenar los huecos.}

\item{combinacionesmaxi}{Cuanto mas grande, mas rapido ira. Problemas de memoria del PC}

\item{verbose}{verbose}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for ..
}
\author{
emilio
}
