% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_recubrimiento.R
\name{obtener_recubrimiento}
\alias{obtener_recubrimiento}
\title{Generate a cover of the first missing places of R}
\usage{
obtener_recubrimiento(R, nhuecos = 1L)
}
\arguments{
\item{R}{}

\item{nhuecos}{Number of (the first missing) places of R to be filled.}
}
\value{
A matrix
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
set.seed(1234)
R <- crear_matriz_aleatoria(ndimension=5L,valordiagonal=10L)
print(R)
cat("Number of missing places: ", sum(is.na(R)),".\n",sep="")
cat("Number of values of R: ", length(obtener_valores_posibles(R)),".\n",sep="")
cat("Unique values of R:", obtener_valores_posibles(R),".\n")
cat("Number of total covers: ", length(obtener_valores_posibles(R)),"^",sum(is.na(R)),".\n",sep="")

## Saca todas las combinaciones rellenando los 4 primeros huecos
mrecubrimiento <- obtener_recubrimiento(R,nhuecos=4L)
head(mrecubrimiento)
dim(mrecubrimiento) # 1296 13
}
\author{
emilio
}
