% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_valores_posibles.R
\name{obtener_valores_posibles}
\alias{obtener_valores_posibles}
\title{Admissible values to fill the missing places.}
\usage{
obtener_valores_posibles(R, solovaloresexistentes = TRUE)
}
\arguments{
\item{R}{}

\item{solovaloresexistentes}{usa solo los valores existentes en R para rellenar el resto de los huecos}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
1
}
\author{
emilio
}
