% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_matrix.R
\name{order_matrix}
\alias{order_matrix}
\title{Get the order of a matrix by columns, selecting some rows.}
\usage{
order_matrix(m, cols = 1L:ncol(m), rows = 1L:nrow(m))
}
\arguments{
\item{m}{Matrix}

\item{cols}{Positions or names of the columns to order.}

\item{rows}{Integer vector with the rows to order.}
}
\value{
An integer vector with the order of the selected rows.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
m <- data.matrix(iris[,1:4])
m[2,1] <- NA
oom <- order_matrix(m,c("Sepal.Length","Sepal.Width"))
head(m[oom,])

rows <- 1:10
oom <- order_matrix(m,c("Sepal.Length","Sepal.Width"),rows)
m[rows,][oo,]

oom <- order_matrix(m,c("Sepal.Length","Sepal.Width"))
oo <- order(m[,"Sepal.Length"],m[,"Sepal.Width"],na.last=FALSE)
all.equal(oom,oo)
}
\author{
emilio
}
