% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacar_nueva_matriz_sin_columnas.R
\name{sacar_nueva_matriz_sin_columnas_o_filas_perdidas}
\alias{sacar_nueva_matriz_sin_columnas_o_filas_perdidas}
\title{.}
\usage{
sacar_nueva_matriz_sin_columnas_o_filas_perdidas(R, verbose = 1L)
}
\arguments{
\item{R}{r}

\item{verbose}{v}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
1
}
\author{
emilio
}
